First release of our TOBA token logo
This logo represent the TOBA volcano that was so fierce that it actually reset the worlds evolution at the time and is now run of the most fertile lands in the world.

Mt Toba Eruption 74,000 years ago. Toba Lake in northern Sumatra is the world's largest active volcanic caldera. The volcanic eruption that resulted in Lake Toba (100 x 30 km) 74,000 years ago, is known to have been by far the biggest eruption of the last 2 million years.

https://en.wikipedia.org/wiki/Toba_catastrophe_theory